+++ aa_forum/templatetags/aa_forum_user.py
@@ -14,9 +14,7 @@ def main_character_name(user: User) -> str:
     """
     Get the users main character name, or return the username if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -35,9 +33,7 @@ def main_character_id(user: User) -> int:
     """
     Get the users main character id, or return 1 if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -54,11 +50,10 @@ def main_character_id(user: User) -> int:
 @register.filter
 def main_character_corporation_name(user: User) -> str:
     """
-    Get the users main character corporation name, or an empty string if no main character
+    Get the users main character corporation name,
+    or an empty string if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -77,9 +72,7 @@ def main_character_corporation_id(user: User) -> int:
     """
     Get the users main character corporation id, or 1 if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -98,9 +91,7 @@ def main_character_alliance_name(user: User) -> str:
     """
     Get the users main character alliance name, or an empty string if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -119,9 +110,7 @@ def main_character_alliance_id(user: User) -> int:
     """
     Get the users main character alliance id, or 1 if no main character
     :param user:
-    :type user:
     :return:
-    :rtype:
     """
 
     if user is None:
@@ -129,6 +118,12 @@ def main_character_alliance_id(user: User) -> int:
 
     try:
         return_value = user.profile.main_character.alliance_id
+
+        # Check if the user is in an alliance
+        try:
+            return_value = int(return_value)
+        except Exception:
+            return_value = 1
     except AttributeError:
         return_value = 1